% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetrics.R
\name{addQCMetricsPerCell_seu}
\alias{addQCMetricsPerCell_seu}
\title{Add cell-wise QC metrics}
\usage{
addQCMetricsPerCell_seu(seu)
}
\arguments{
\item{seu}{A Seurat object, can contain a assay of any name, such as 
"RNA" for chromium, "Spatial" for Visium, or "Xenium" for Xenium data.}
}
\value{
Returns a Seurat object with per-cell flags.
Two per-cell flags are derived.  `libsize_drop` and `mito_drop` are stored 
in column meta data as per-cell low total count flag and high mitochondria  
gene percentage flag, respectively. Column-wise meta data of a Seurat object    
can be retrieved by `seu[["libsize_drop"]]` or `seu$libsize_drop`.
}
\description{
Inspired by OSCA book, the following function automatically adds cell-wise 
QC metrics on low library size and high mitochondria percentage to a Seurat 
object and store for later use.
}
\examples{
\dontrun{
## Chromium Example
data("pbmc_small")
pbmc_small <- addQCMetricsPerCell_seu(pbmc_small)

## Visium Example
data_dir <- '/path/to/Visium/outs/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
vis <- Seurat::Load10X_Spatial(data.dir = data_dir)
vis <- SeuratQCUtils::addQCMetricsPerCell_seu(vis)

## Xenium Example
data_dir <- '/path/to/Xenium/outs/'
list.files(data_dir) # Should show cell_feature_matrix.h5
xe <- Seurat::LoadXenium(data.dir = data_dir)
xe <- SeuratQCUtils::addQCMetricsPerCell_seu(xe)
}
}
