% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetrics.R
\name{addQCMetricsPerGene_seu}
\alias{addQCMetricsPerGene_seu}
\title{Inspired by OSCA book, the following function automatically adds gene-wise  
QC metric on low abundance genes to a Seurat object and store for later use.}
\usage{
addQCMetricsPerGene_seu(seu)
}
\arguments{
\item{seu}{A Seurat object, can contain a assay of any name, such as 
"RNA" for chromium, "Spatial" for Visium, or "Xenium" for Xenium data.}
}
\value{
Returns a Seurat object with both per-cell and per-gene flags.
2. Two per-gene statistics and flag are derived. `means` and `lowgenecount_drop`
are stored in gene-wise meta data, which can be retrieved by, e.g. for a raw
assay with name "RNA", `seu[["RNA"]][["means"]]`.
}
\description{
Inspired by OSCA book, the following function automatically adds gene-wise  
QC metric on low abundance genes to a Seurat object and store for later use.
}
\examples{
\dontrun{
data("pbmc_small")
pbmc_small <- addQCMetricsPerGene_seu(pbmc_small)

## Visium Example
data_dir <- '/path/to/Visium/outs/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
vis <- Seurat::Load10X_Spatial(data.dir = data_dir)
vis <- SeuratQCUtils::addQCMetricsPerGene_seu(vis)

## Xenium Example
data_dir <- '/path/to/Xenium/outs/'
list.files(data_dir) # Should show cell_feature_matrix.h5
xe <- Seurat::LoadXenium(data.dir = data_dir)
xe <- SeuratQCUtils::addQCMetricsPerGene_seu(xe)
}
}
