% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistograms.R
\name{plot_Hist_High_Mito_Props}
\alias{plot_Hist_High_Mito_Props}
\title{A per-cell mitochondria percentage QC histogram that takes a Seurat object   
that has been preprocessed with `SeuratQCUtils::addQCMetrics()`, such that  
flag `mito_drop` and continuous variable `percent.mt` exist in meta.data.   
This function sources helper function `SeuratQCUtils::plotHist()` to 
generate histogram.}
\usage{
plot_Hist_High_Mito_Props(seu, yvar = "mito_drop")
}
\arguments{
\item{seu}{A Seurat object with per-cell QC flag `mito_drop` and `percent.mt`
stored in meta.data.}

\item{yvar}{A binary flag name `mito_drop` with level TRUE/FALSE
indicating if a cell/spot will be removed.}
}
\value{
A histogram, colored by QC `mito_drop` flag status.
}
\description{
A per-cell mitochondria percentage QC histogram that takes a Seurat object   
that has been preprocessed with `SeuratQCUtils::addQCMetrics()`, such that  
flag `mito_drop` and continuous variable `percent.mt` exist in meta.data.   
This function sources helper function `SeuratQCUtils::plotHist()` to 
generate histogram.
}
\examples{
\dontrun{
## Chromium Example
data_dir <- '/path/to/Chromium/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
chrom <- Seurat::Read10X_h5(paste0(data_dir, "filtered_feature_bc_matrix.h5"))
chrom <- CreateSeuratObject(counts = chrom)
chrom <- SeuratQCUtils::addQCMetricsPerCell_seu(chrom)
if(any(chrom$mito_drop)){
  p <- SeuratQCUtils::plot_Hist_High_Mito_Props(chrom, yvar = "mito_drop")
}
p

## Visium Example
data_dir <- '/path/to/Visium/outs/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
vis <- Seurat::Load10X_Spatial(data.dir = data_dir)
vis <- SeuratQCUtils::addQCMetricsPerCell_seu(vis)
if(any(vis$mito_drop)){
  p <- SeuratQCUtils::plot_Hist_High_Mito_Props(vis, yvar = "mito_drop")
}
p

## Xenium Example (Xenium has no mitochondria genes)

}
}
