% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistograms.R
\name{plot_Hist_Low_Abun_Genes}
\alias{plot_Hist_Low_Abun_Genes}
\title{A per-gene abundance QC histogram}
\usage{
plot_Hist_Low_Abun_Genes(seu)
}
\arguments{
\item{seu}{A Seurat object with per-gene QC flag `lowgenecount_drop` and 
`means` stored in row-wise meta.data.}
}
\value{
A histogram, colored by QC `lowgenecount_drop` flag status.
}
\description{
A per-gene abundance QC histogram that takes a Seurat object that has been 
preprocessed with `SeuratQCUtils::addQCMetrics()`, such that flag 
`lowgenecount_drop` and continuous variable gene `means` exist in row-wise
meta.data. This function sources helper function `SeuratQCUtils::plotHist()` 
to generate histogram.
}
\examples{
\dontrun{
## Chromium Example
data_dir <- '/path/to/Chromium/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
chrom <- Seurat::Read10X_h5(paste0(data_dir, "filtered_feature_bc_matrix.h5"))
chrom <- CreateSeuratObject(counts = chrom)
chrom <- SeuratQCUtils::addQCMetricsPerGene_seu(chrom)
if(any(unlist(chrom[["RNA"]][["lowgenecount_drop"]]))){
  p <- SeuratQCUtils::plot_Hist_Low_Abun_Genes(chrom)
}
p

## Visium Example
data_dir <- '/path/to/Visium/outs/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
vis <- Seurat::Load10X_Spatial(data.dir = data_dir)
vis <- SeuratQCUtils::addQCMetricsPerGene_seu(vis)
if(any(unlist(vis[["Spatial"]][["lowgenecount_drop"]]))){
  p <- SeuratQCUtils::plot_Hist_Low_Abun_Genes(vis)
}
p
## Xenium Example
data_dir <- '/path/to/Xenium/outs/'
list.files(data_dir) # Should show cell_feature_matrix.h5
xe <- Seurat::LoadXenium(data.dir = data_dir)
xe <- SeuratQCUtils::addQCMetricsPerGene_seu(xe)
if(any(unlist(xe[["Xenium"]][["lowgenecount_drop"]]))){
  p <- SeuratQCUtils::plot_Hist_Low_Abun_Genes(xe)
}
p
}
}
