% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistograms.R
\name{plot_Hist_Low_Lib_Sizes}
\alias{plot_Hist_Low_Lib_Sizes}
\title{A per-cell library size QC histogram}
\usage{
plot_Hist_Low_Lib_Sizes(seu, yvar = "libsize_drop")
}
\arguments{
\item{seu}{A Seurat object with per-cell QC flag `libsize_drop` stored in 
meta.data}

\item{yvar}{A binary flag name `libsize_drop` with level TRUE/FALSE
indicating if a cell/spot will be removed.}
}
\value{
A histogram, colored by QC `libsize_drop` flag status.
}
\description{
A per-cell library size QC histogram that takes a Seurat object that has  
been preprocessed with `SeuratQCUtils::addQCMetrics()`, such that flag 
`libsize_drop` exists in meta.data. This function sources helper function to 
generate histogram with `SeuratQCUtils::plotHist()`.
}
\examples{
\dontrun{
## Chromium Example
data_dir <- '/path/to/Chromium/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
chrom <- Seurat::Read10X_h5(paste0(data_dir, "filtered_feature_bc_matrix.h5"))
chrom <- CreateSeuratObject(counts = chrom)
chrom <- SeuratQCUtils::addQCMetricsPerCell_seu(chrom)
if(any(chrom$libsize_drop)){
  p <- SeuratQCUtils::plot_Hist_Low_Lib_Sizes(chrom, yvar = "libsize_drop")
}
p

## Visium Example
data_dir <- '/path/to/Visium/outs/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
vis <- Seurat::Load10X_Spatial(data.dir = data_dir)
vis <- SeuratQCUtils::addQCMetricsPerCell_seu(vis)
if(any(vis$libsize_drop)){
  p <- SeuratQCUtils::plot_Hist_Low_Lib_Sizes(vis, yvar = "libsize_drop")
}
p

## Xenium Example
data_dir <- '/path/to/Xenium/outs/'
list.files(data_dir) # Should show cell_feature_matrix.h5
xe <- Seurat::LoadXenium(data.dir = data_dir)
xe <- SeuratQCUtils::addQCMetricsPerCell_seu(xe)
if(any(xe$libsize_drop)){
  p <- SeuratQCUtils::plot_Hist_Low_Lib_Sizes(xe, yvar = "libsize_drop")
}
p

}
}
