% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R
\name{plotBar_seu}
\alias{plotBar_seu}
\title{Cell-type frequency bar plot}
\usage{
plotBar_seu(
  seu,
  decreasing = TRUE,
  ylabel = "Percentage",
  title = "Cell-type Frequency in Annotated Chromium"
)
}
\arguments{
\item{seu}{A Seurat object contains "Annotation" column in the meta data}

\item{decreasing}{logical that indicates if the bars should be arranged 
decreasing or not}

\item{ylabel}{default "Percentage", as we are plotting cell-type proportion}

\item{title}{default "Cell-type Frequency in Annotated Chromium"}
}
\value{
Returns a bar plot with cell-type percentages arranged in decreasing 
order for the input annotated Seurat object.
}
\description{
Create cell-type frequency bar plot in annotated Seurat object
}
\examples{
\dontrun{
data("pbmc_small")
pbmc_small$Annotation <- pbmc_small$groups
p <- SeuratQCUtils::plotBar_seu(pbmc_small)
p
}
}
