% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistograms.R
\name{plotHist}
\alias{plotHist}
\title{A generic histogram plotting function that will be used for gene-wise and 
cell-wise QC flags on their corresponding continuous variables.}
\usage{
plotHist(plot_df, xvar = "logtotal", yvar = "libsize_drop")
}
\arguments{
\item{plot_df}{A data frame contains one continuous variable and one binary 
flag for indicating which cells/spots/genes did not pass QC.}

\item{xvar}{One continuous variable for plotting, e.g. log of total library
size.}

\item{yvar}{One binary flag with level TRUE/FALSE indicating if a cell/spot 
or a gene, e.g. libsize_drop.}
}
\value{
A histogram, colored by QC flag status.
}
\description{
A generic histogram plotting function that will be used for gene-wise and 
cell-wise QC flags on their corresponding continuous variables.
}
\examples{
\dontrun{
## Chromium Example 
data("pbmc_small")
pbmc_small <- SeuratQCUtils::addQCMetricsPerCell_seu(pbmc_small)
plot_df <- data.frame(logtotal = log(as.numeric(pbmc_small$nCount_RNA)),
                      pbmc_small@meta.data)
p <- plotHist(plot_df, xvar = "logtotal", yvar = "libsize_drop")
p

## Visium Example
data_dir <- '/path/to/Visium/outs/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
vis <- Seurat::Load10X_Spatial(data.dir = data_dir)
vis <- SeuratQCUtils::addQCMetricsPerCell_seu(vis)
plot_df <- data.frame(logtotal = log(as.numeric(vis$nCount_Spatial)),
                      vis@meta.data)
p <- plotHist(plot_df, xvar = "logtotal", yvar = "libsize_drop")
p

## Xenium Example
data_dir <- '/path/to/Xenium/outs/'
list.files(data_dir) # Should show cell_feature_matrix.h5
xe <- Seurat::LoadXenium(data.dir = data_dir)
xe <- SeuratQCUtils::addQCMetricsPerCell_seu(xe)
p <- SeuratQCUtils::plotHist(plot_df, xvar = "logtotal", yvar = "libsize_drop")
p

}
}
