% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCSpatial.R
\name{plotQCSpatial_seu}
\alias{plotQCSpatial_seu}
\title{Inspired by ggspavis in OSTA book, spatial QC plot for binary flags, such 
as `libsize_drop`, to see which spots or cells have not pass QC check in 
`addQCMetrics_seu()`.}
\usage{
plotQCSpatial_seu(seu_sp, flag = "libsize_drop")
}
\arguments{
\item{seu_sp}{A Visium or Xenium Seurat object with spatial coordinates.}

\item{flag}{A binary flag contains two classes of TRUE or FALSE}
}
\value{
A spatial QC plot highlighting spots to be removed after QC in red.
}
\description{
Inspired by ggspavis in OSTA book, spatial QC plot for binary flags, such 
as `libsize_drop`, to see which spots or cells have not pass QC check in 
`addQCMetrics_seu()`.
}
\examples{
\dontrun{
## Visium Example
data_dir <- '/path/to/Visium/outs/'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
vis <- Seurat::Load10X_Spatial(data.dir = data_dir)
vis <- SeuratQCUtils::addQCMetricsPerCell_seu(vis)
p <- SeuratQCUtils::plotQCSpatial_seu(seu_sp = vis, flag = "libsize_drop")
p

## Xenium Example
data_dir <- '/path/to/Xenium/outs/'
list.files(data_dir) # Should show cell_feature_matrix.h5
xe <- Seurat::LoadXenium(data.dir = data_dir)
xe <- SeuratQCUtils::addQCMetricsPerCell_seu(xe)
p <- SeuratQCUtils::plotQCSpatial_seu(seu_sp = xe, flag = "libsize_drop")
p
}
}
